Fixed issues for next stable release
git-svn-id: https://ekptrac.physik.uni-karlsruhe.de/svn/grid-control/trunk/grid-control@1699 29098800-e21f-4f28-a8ff-84ee76f0ac90

+++ packages/grid_control_gui/report_basic.py
@@ -39,7 +39,7 @@ class BasicProgressBar(object):
 
 		# Print percentage
 		text = str(done) + '%'
-		textPos = (self._width - len(text) + 1) / 2
+		textPos = int((self._width - len(text) + 1) / 2)
 		self._bar = self._bar[0:textPos] + text + self._bar[textPos+len(text):]
 
 	def __str__(self):
